Tests/LibWeb/Text/expected/css/update-pseudo-elements-on-hover.txt

Not hovering: 16
Hovering: 78
Not hovering: 16
